Move all theme-related logic into a ThemeWrapper
